ambiguous std::max on windows

+++ pol-core/pol/module/utilmod.cpp
@@ -222,7 +222,7 @@ Bscript::BObjectImp* UtilExecutorModule::mf_StrFormatTime()
   }
 
   std::string buffer;
-  buffer.reserve( std::max( format_string->length() * 2, 50ul ) );
+  buffer.reserve( std::max( format_string->length() * 2, (size_t)50u ) );
   while ( strftime( buffer.data(), buffer.capacity(), format_string->data(), &time_struct ) == 0 )
   {
     buffer.reserve( buffer.capacity() * 2 );
